@@ -1,11 +1,15 @@
 package com.sap.cloud.extensibility.services;
 
+import javax.ejb.Stateless;
+
+import org.apache.commons.configuration.PropertiesConfiguration;
 import org.slf4j.Logger;
 
 import com.sap.cloud.extensibility.model.OneTimeCustomerOrder;
 import com.sap.cloud.sdk.cloudplatform.logging.CloudLoggerFactory;
 import com.sap.cloud.sdk.odatav2.connectivity.ODataException;
 
+@Stateless
 public class OneTimeCustomerRecordService {
 
 	private static final Logger LOGGER = CloudLoggerFactory.getLogger(OneTimeCustomerRecordService.class);
@@ -21,15 +25,19 @@ public class OneTimeCustomerRecordService {
 	 * @throws Exception
 	 */
 	public OneTimeCustomerOrder create(OneTimeCustomerOrder otcOrder) throws Exception {
-		
+
 		OneTimeCustomerOrder otcOrderCreated = null;
-		
+
+		PropertiesConfiguration config = new PropertiesConfiguration();
+
+		config.load("application.properties");
+
 		try {
 
 			if (otcOrder != null) {
 
 				OneTimeCustomerRecordCreateFluentHelper orderHelper = new OneTimeCustomerRecordCreateFluentHelper(
-						otcOrder);
+						config.getString("s4cld.onetimecustomerrecord_servicepath"), otcOrder);
 
 				otcOrderCreated = orderHelper.execute();
 
@@ -38,9 +46,9 @@ public class OneTimeCustomerRecordService {
 			}
 
 		} catch (ODataException e) {
-			
+
 			LOGGER.error("Exception occured while creating One Time Customer Record :: ", e);
-			
+
 			throw new Exception(e);
 		}
 
