@@ -3,11 +3,11 @@ package com.sap.cloud.extensibility.services;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.ejb.Stateless;
 import javax.inject.Inject;
 
 import org.slf4j.Logger;
 
-import com.sap.cloud.extensibility.model.CustomProduct;
 import com.sap.cloud.sdk.cloudplatform.logging.CloudLoggerFactory;
 import com.sap.cloud.sdk.odatav2.connectivity.ODataException;
 import com.sap.cloud.sdk.s4hana.datamodel.odata.namespaces.productmaster.Product;
@@ -16,6 +16,8 @@ import com.sap.cloud.sdk.s4hana.datamodel.odata.services.DefaultProductMasterSer
 /**
  * The Class ProductService.
  */
+
+@Stateless
 public class ProductService {
 
 	/** The default product master service. */
@@ -79,33 +81,30 @@ public class ProductService {
 		return productsList;
 
 	}
-	
-	public List<CustomProduct> findCustomerProductsByProductGroup(String productGroup) throws Exception {
+
+	public List<Product> findCustomerProductsByProductGroup(String productGroup) throws Exception {
 
 		List<Product> productsList = null;
-		
-		List<CustomProduct> customProductsList = new ArrayList<CustomProduct>();
-		
+
+		List<Product> customProductsList = new ArrayList<Product>();
+
 		try {
 
 			logger.info("productGroup :: " + productGroup);
 
-			productsList = defaultProductMasterService.getAllProduct().filter(Product.PRODUCT_GROUP.eq(productGroup)).
-					execute();
-			
-			logger.info("productsList size::"+productsList.size());
-			
-			for(Product p: productsList ) {
-				
-				logger.info("YY1_WebSaleable_PRD"+p.getCustomField("YY1_SaleableProduct_PRD"));
-				
-				customProductsList.add(CustomProduct.of(p));
-				
+			productsList = defaultProductMasterService.getAllProduct().filter(Product.PRODUCT_GROUP.eq(productGroup))
+					.execute();
+
+			logger.info("productsList size::" + productsList.size());
+
+			for (Product p : productsList) {
+
+				logger.info("YY1_WebSaleable_PRD" + p.getCustomField("YY1_SaleableProduct_PRD"));
+
+				customProductsList.add(p);
+
 			}
-			for(CustomProduct cp: customProductsList) {
-				logger.info("cp websaleble::"+cp.getCustomWebSaleble());
-			}
-			
+
 			logger.info("customProductsList :: " + customProductsList);
 
 		} catch (ODataException e) {
@@ -117,6 +116,5 @@ public class ProductService {
 		return customProductsList;
 
 	}
-
 
 }
